chore: move market filtering to MarketContent
