+++ src/components/ui/lending/MarketContent/MarketContent.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import React, { useState } from "react";
+import React, { useState, useMemo } from "react";
 import MarketCard from "@/components/ui/lending/MarketContent/MarketCard";
 import CardsList from "@/components/ui/CardsList";
 import {
@@ -11,6 +11,7 @@ import {
   UserSupplyPosition,
 } from "@/types/aave";
 import { TokenTransferState } from "@/types/web3";
+import { LendingFilters, LendingSortConfig } from "@/types/lending";
 
 const ITEMS_PER_PAGE = 10;
 
@@ -21,6 +22,8 @@ interface MarketContentProps {
   tokenTransferState: TokenTransferState;
   onSupply: (market: UnifiedMarketData) => void;
   onBorrow: (market: UnifiedMarketData) => void;
+  filters: LendingFilters;
+  sortConfig: LendingSortConfig | null;
 }
 
 const MarketContent: React.FC<MarketContentProps> = ({
@@ -28,10 +31,67 @@ const MarketContent: React.FC<MarketContentProps> = ({
   tokenTransferState,
   onSupply,
   onBorrow,
+  filters,
+  sortConfig,
 }) => {
   const [currentPage, setCurrentPage] = useState(1);
 
-  if (!unifiedMarkets || unifiedMarkets.length === 0) {
+  // Apply filtering and sorting to unified markets
+  const filteredAndSortedMarkets = useMemo(() => {
+    if (!unifiedMarkets) return null;
+
+    let filtered = unifiedMarkets;
+
+    // Filter by asset
+    if (filters.assetFilter) {
+      const filterLower = filters.assetFilter.toLowerCase();
+      filtered = unifiedMarkets.filter((market) => {
+        return (
+          market.underlyingToken.symbol.toLowerCase().includes(filterLower) ||
+          market.underlyingToken.name.toLowerCase().includes(filterLower) ||
+          market.marketName.toLowerCase().includes(filterLower)
+        );
+      });
+    }
+
+    // Sort unified markets
+    if (sortConfig) {
+      filtered = [...filtered].sort((a, b) => {
+        let aValue: number;
+        let bValue: number;
+
+        switch (sortConfig.column) {
+          case "supplyApy":
+            aValue = a.supplyData.apy;
+            bValue = b.supplyData.apy;
+            break;
+          case "borrowApy":
+            aValue = a.borrowData.apy;
+            bValue = b.borrowData.apy;
+            break;
+          case "suppliedMarketCap":
+            aValue = a.supplyData.totalSuppliedUsd;
+            bValue = b.supplyData.totalSuppliedUsd;
+            break;
+          case "borrowedMarketCap":
+            aValue = a.borrowData.totalBorrowedUsd;
+            bValue = b.borrowData.totalBorrowedUsd;
+            break;
+          default:
+            aValue = 0;
+            bValue = 0;
+        }
+
+        return sortConfig.direction === "asc"
+          ? aValue - bValue
+          : bValue - aValue;
+      });
+    }
+
+    return filtered;
+  }, [unifiedMarkets, filters, sortConfig]);
+
+  if (!filteredAndSortedMarkets || filteredAndSortedMarkets.length === 0) {
     return (
       <div className="text-center py-16">
         <div className="text-[#A1A1AA]">no markets found</div>
@@ -39,8 +99,10 @@ const MarketContent: React.FC<MarketContentProps> = ({
     );
   }
 
-  const totalPages = Math.ceil(unifiedMarkets.length / ITEMS_PER_PAGE);
-  const paginatedMarkets = unifiedMarkets.slice(
+  const totalPages = Math.ceil(
+    filteredAndSortedMarkets.length / ITEMS_PER_PAGE,
+  );
+  const paginatedMarkets = filteredAndSortedMarkets.slice(
     (currentPage - 1) * ITEMS_PER_PAGE,
     currentPage * ITEMS_PER_PAGE,
   );
@@ -72,7 +134,7 @@ const MarketContent: React.FC<MarketContentProps> = ({
       totalPages={totalPages}
       onPageChange={handlePageChange}
       itemsPerPage={ITEMS_PER_PAGE}
-      totalItems={unifiedMarkets.length}
+      totalItems={filteredAndSortedMarkets.length}
     />
   );
 };
